move search to components
